% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_linear_regression_ME}
\alias{fit_linear_regression_ME}
\title{Fit and summarize the Linear Regression with Measurement Error Stan model}
\usage{
fit_linear_regression_ME(N, x_meas, y, a, tau)
}
\arguments{
\item{N}{Number of observations.}

\item{x_meas}{Measured independent variable.}

\item{y}{Dependent variable.}

\item{a}{Coefficient parameter.}

\item{tau}{Tau parameter.}
}
\value{
Summary of the fit.
}
\description{
This function fits the Linear Regression with Measurement Error Stan model and returns the summary.
}
