% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_logistic_regression_ME}
\alias{fit_logistic_regression_ME}
\title{Fit and summarize the Logistic Regression with Measurement Error Stan model}
\usage{
fit_logistic_regression_ME(model, N, D, y, x_meas, x_in, a, tau)
}
\arguments{
\item{N}{Number of observations.}

\item{D}{Number of covariates, excluding the column with measurement noise}

\item{y}{Dependent variable.}

\item{x_meas}{Input vector with measurement noise.}

\item{x_in}{Input matrix x_in for the additional D covariates, with dimensions N rows by D columns}

\item{a}{Coefficient parameter.}

\item{tau}{Tau parameter.}
}
\value{
Summary of the fit.
}
\description{
This function fits the Logistic Regression with Measurement Error Stan model and returns the summary.
}
