% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_tau}
\alias{fit_tau}
\title{Fit and summarize the Tau Stan model}
\usage{
fit_tau(model, N, Z, Zp, a)
}
\arguments{
\item{model}{Compiled Stan model for Tau.}

\item{N}{Number of observations.}

\item{a}{Coefficient parameter.}

\item{Y}{Dependent variable.}

\item{Yp}{Measured dependent variable.}
}
\value{
Summary of the fit.
}
\description{
This function fits the provided Tau Stan model and returns the summary.
}
