% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_stan_models}
\alias{load_stan_models}
\title{Load and compile Stan models}
\usage{
load_stan_models()
}
\value{
A list of compiled Stan models.
}
\description{
This function loads and compiles the Stan models.
}
