% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_linear_regression}
\alias{fit_linear_regression}
\title{Fit and summarize the Linear Regression Stan model}
\usage{
fit_linear_regression(model, N, D, x, y)
}
\arguments{
\item{model}{Compiled Stan model for Linear Regression.}

\item{N}{Number of observations.}

\item{D}{Number of covariates.}

\item{x}{Independent variable.}

\item{y}{Dependent variable.}
}
\value{
Summary of the fit.
}
\description{
This function fits the provided Linear Regression Stan model and returns the summary.
}
